Fix bug related to UTC vs. EST.
The regulations.gov API expects dates to be *sent* in EST (with
DST) but *returns* dates in UTC.  This causes a 4 or 5 hour
difference in time, resulting in missing data when the work
generator creates new searches.

This fix converts all dates to UTC with the conversion to ETC
at the last moment before the request is sent to the server.