Make work server take more recent jobs from right of job waiting queue
The job_queue adds jobs to the front of the job waiting queue. This causes older
jobs to be at the front of this list. The work server should pull the newest jobs,
meaning the jobs later in the queue. This will allow the software to download the
current rules, as requested.