Add discussion of PYTHONUNBUFFERED=TRUE in .env files
When looking at the docker-compose logs, we were seeing bursts of log
activity.  This was caused because Python was buffering output - from
Docker's perspective, there were no logs, and then a large number of
lines would arrive in a fraction of a second.

By setting the PYTHONUNBUFFERED environment variable, Python
will stop using buffered output, and instead generate output
immediately.  This causes the lines to arrive at Docker as soon
as they are generated, which makes our logs accurately reflect
what is happening in the containers.