Add code to interact with RabbitMQ
This commit adds a class that will replace the jobs_waiting_queue in
Redis.  The Message Queue (MQ) has the same time, but stores data
more efficiently than Redis.

The RabbitMQ class encapsulates all interactions with RabbitMQ.
This code cannot be easily tested, so the "test" included in this
commit mocks everything out - just to ensure we achieve our
coverage metric.

The Dockerfile defines a container based on RabbitMQ's pre-defined
container.  Ours adds curl, which will be used in a health check
(see future commit on docker-compose.yml).

Finally, this commit includes a MockRabbit class that implements
the methods of the RabbitMQ class.  This will be used for testing
when RabbitMQ is integrated into the system.