Add RabbitMQ and health checks for all DB containers
This version of the compose file adds health checks for
mongo, redis, and rabbitmq.  This allows docker to determine
if these containers are up-and-running.  Additionally, it
specifies that the work_generator, work_server, and dashboard
depend on the DB containers.  As a result, when we run
docker-compose up, the DBs will be started first, and the
work_generator, work_server, and dashboard will not start
until the 3 DB are healthy.  This solves the problem that
portions of the system crashed when they expected the DB
to be ready.