Replace Redis with RabbitMQ for jobs_waiting_queue
The JobQueue now interacts with both Redis and RabbitMQ.  RabbitMQ
holds the jobs_waiting_queue, but all other functionality remains
in Redis (in progress, counts, etc.).

This commit also adds a method get_job_stats that returns statistical
information about jobs.  This will be used in the dashboard.

The tests use the MockRabbit class to avoid interactions with a
real rabbitmq instance.  I don't like that the mocking of Redis
and RabbitMQ are different.  the Redis() constructor connects
immediately, so that has to be mocked before the JobQueue class
is created.  RabbitMQ does not create a connection in the constrcutor,
so we can replace the instance after the constructor runs.

In the future, I'd like to fix this inconsistency.  But doing so
now would require adding a second parameter to the JobQueue class.
Because JobQueue is used through the work generator, work server,
and dashboard, this would have been a significantly larger change.