make the work generator look for duplication on disk instead of mongo.
The work generator checks to see if the data exists before it adds it
to the queue.  Previously, this looked at Mongo, which was taking
.5 to 1.5 seconds per check.  This made the work generation process
too slow - many clients were sitting idle.

This change uses the PathGenerator to compute the path of the resulting
file.  Because search results have id, agency, etc., they are close
enough that the PathGenerator can compute the path (with a small hack).

This method also allowed us to get rid of the DataStorage class in
the work generation process.  This simplifies the interface to these
objects.