Add try except around the text extraction call
The except is currently only catching a generic exception. This is because
the error that was thrown in production was a ValueError and happened as a
result of an internal error that we have no control over. By the same
token, it could be some other general error. I think it is best to just
use a generic catch.
Lines starting