Improve Exception Handling
Removed JobQueueException as JobQueue is never actually
throwing it.
Added APITimeout Exception instead of letting
requests throw its default exception.
Removed everything to do with storing errors in
an 'error' key. This was a quirk of the work server using
HTTP and returning errors through an 'error' key in the JSON.
Basic common sense that we raise exceptions when errors happen
in Python.